Move GroupConfig from main to group_config
The main module has too much logic that's hard to reuse by other
modules, GroupConfig is the entry point for any other program attempting
to reuse charmhub-lp-tools as a library, making it live within its own
module we lower the barrier of adoption.
